requirements.txt
flask
google-generativeai
python-Levenshtein
fuzzywuzzy